extract state base
