+++ addon/document/-document-state-mixin.js
@@ -1,0 +1,4 @@
+import { keys } from './internal/-document-state';
+import create from './-create-state-mixin';
+
+export default create(keys);
